+++ llvm-hs/src/Control/Monad/Trans/AnyCont.hs
@@ -1,5 +1,5 @@
 {-# LANGUAGE
-  RankNTypes
+  CPP, RankNTypes
   #-}
 module Control.Monad.Trans.AnyCont where
 
@@ -21,7 +21,9 @@ instance Applicative (AnyContT m) where
 instance Monad m => Monad (AnyContT m) where
   AnyContT f >>= k = AnyContT $ f >>= unAnyContT . k
   return a = AnyContT $ return a
+#if !(MIN_VERSION_base(4,13,0))
   fail s = AnyContT (ContT (\_ -> Cont.fail s))
+#endif
 
 instance MonadFail m => MonadFail (AnyContT m) where
   fail s = AnyContT (ContT (\_ -> Fail.fail s))