+++ llvm-hs-pure/CHANGELOG.md
@@ -1,3 +1,8 @@
+## unreleased
+
+* The functions in `LLVM.IRBuilder.Constant` no longer return a
+  monadic context. To recover the previous behavior use `pure`.
+
 ## 8.0.0 (2019-03-10)
 
 * Upgrade to LLVM 8